optimize vault name path, organize config classes
